pyproject.toml
[tool.poetry]
name = "wowhead-llm"
version = "0.1.0"
description = "Uses Mistral AI to query documents"
authors = ["Ryan Davis <ryancraigdavis@github.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
llama-index-core = "^0.10.3"
llama-index-readers-file = "^0.1.3"
llama-index-llms-ollama = "^0.1.1"
llama-index-embeddings-huggingface = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
